feat: rename templates for follow the pattern
