+++ frontend/web/src/lib/components/templates/admin/BoardGamesRegister.svelte
