+++ frontend/web/src/lib/components/templates/admin/BoardGamesUpdate.svelte
