+++ frontend/web/src/lib/components/templates/admin/EventDetails.svelte
